[refactor] 패키징 추가
패키징 추가, companion object 삭제, 카드 뷰 스크롤 순서 수정

Related to : #9, #10
