+++ app/src/main/java/meaning/morning/data/GroupData.kt
@@ -1,4 +1,4 @@
-package meaning.morning
+package meaning.morning.data
 
 data class GroupData(
     val groupName: String,
