+++ app/src/main/java/meaning/morning/presentation/HomeFragment.kt
@@ -1,10 +1,11 @@
-package meaning.morning
+package meaning.morning.presentation
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import meaning.morning.R
 
 
 class HomeFragment : Fragment() {
