+++ app/src/main/res/layout/item_group_recommend.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="recommendGroup"
-            type="meaning.morning.RecommendGroupData" />
+            type="meaning.morning.data.RecommendGroupData" />
     </data>
 
     <RelativeLayout
